+++ src/win32/nt4_dev.h
@@ -36,12 +36,12 @@ BOOL NT4CloseDrive ( const HANDLE hDrv );
 
 BOOL NT4ReadSector ( const HANDLE hDrv,
                      const long   iSect,
-                     const int    iSize,
+                     const size_t iSize,
                      void * const lpvoidBuf );
 
 BOOL NT4WriteSector ( const HANDLE hDrv,
                       const long   iSect,
-                      const int    iSize,
+                      const size_t iSize,
                       const void * const lpvoidBuf );
 
 ULONG NT4GetDriveSize ( const HANDLE hDrv );
